Adding graphical elements
I added a simple class for a button, all it can do is check if it's
pressed or not but hey that's exactly what buttons do all the time :joy:
so that sounds pretty good. You can also assign a sprite to it.

Oh and I stole InputHelper from our tetris project