Fix GuiList buttons not disappearing
GuiList buttons would persist when all labels were empty; now, there is a visibility flag in GuiElement which makes a draw conditional.
When the visibility of a GuiLabel is changed, it notifies its parent of this, if their parent is a GuiList. When no labels in the list are visible, the list is no longer drawn, causing the buttons to no longer be drawn.